backend/requirements.txt
alembic==1.10.3
anyio==3.6.2
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
emails==0.6
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.10
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
premailer==3.10.0
pyasn1==0.5.0
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
python-jenkins==1.7.0
